Random question generation
problem.py: Is a class that aggregates the question we are asking the
player and all the correct and wrong answers, as two different
attributes, and some other important information (see documentation).

ourmath.py: Is a collection of functions that return a Problem object
with the numbers we want to fill in the board and some extra
information (see documentation).

boardexample.py: It's an example on how to use the problems generator,
the Problem object and so on… So you can use it to populate the board.