Added Images and functionality to main char
I added a lot of images, such as different directions for the main
characters more scenes and the first two levels of our games.
Everything is under the images folder. I also added the functionality
that the character will now face the direction that he is traveling. I
plan on making it so when a user presses left or right he only travels
a set distance.