Added Battery Status info and upgrade Player
The game now has an ammo limit so the user can not shoot unlimited
lasers. The user can only shoot three times and each time will lose a
battery which is the ammo. Also added a empty gun sound when the user
runs out of ammo. Also added a forcefield powering down sound.