Ourmath2 class improved
Now I think we have good questions being generated a way more often
than previously. I believe it matches with 5th graders, by comparing
with the math online game we found. Stage 3 fractions might still be
generated a bit hard sometimes, but I believe that it will be
challenging enough most of the times.
Please try to see and tell me what you guys think. Remember it's
random… It might happen to get an easy number in a difficulty level
that was supposed to be "hard".