Avoidable numbers
If you don't want to generate a second stage asking the same question
from the previous stage, you can now tell the math class that you want
avoid choosing that question… So the problem will never be repeated in
the same planet, if we want.
To use this optional argument, just call the method
generatesMultiplesProblems(27,1,avoid)… with avoid being an integer
that you don't want to be pick.
the same works for generatesMultiplesEqualities(27,1,avoid)…
Fractions method does not have such feature YET.